feat: 유저 엔티티 기록 값 변경 로직 수정, API 추가
- 퀘스트 등록수 증가 시 마지막 등록일 갱신
- 퀘스트 완료수 증가 시 마지막 완료일 갱신
- 경험치 골드 획득 시 골드 획득량 갱신
- 골드 획득, 사용량 타입 Long으로 변경
- 업적에서 모든 기록 값을 수용하기 위해 현재 값 타입을 Long으로 변경

+++ achievement/src/main/kotlin/dailyquest/achievement/dto/AchievementAchieveRequest.kt
@@ -5,12 +5,12 @@ import dailyquest.achievement.entity.AchievementType
 data class AchievementAchieveRequest(
     val type: AchievementType,
     val userId: Long,
-    val currentValue: Int
+    val currentValue: Long
 ) {
 
     companion object {
         @JvmStatic
-        fun of(type: AchievementType, userId: Long, currentValue: Int): AchievementAchieveRequest {
+        fun of(type: AchievementType, userId: Long, currentValue: Long): AchievementAchieveRequest {
             return AchievementAchieveRequest(type, userId, currentValue)
         }
     }
